+++ Tests/KeyboardKitTests/_Keyboard/KeyboardInputViewControllerTests.swift
@@ -127,7 +127,7 @@ class KeyboardInputViewControllerTests: XCTestCase {
         XCTAssertNotNil(vc.services.calloutActionProvider as? Callouts.StandardActionProvider)
         XCTAssertNotNil(vc.services.dictationService as? Dictation.DisabledKeyboardService)
         XCTAssertNotNil(vc.services.keyboardBehavior as? Keyboard.StandardBehavior)
-        XCTAssertNotNil(vc.services.layoutProvider as? StandardKeyboardLayoutProvider)
+        XCTAssertNotNil(vc.services.layoutProvider as? KeyboardLayout.StandardProvider)
         XCTAssertNotNil(vc.services.styleProvider as? StandardKeyboardStyleProvider)
     }